Move Fullscreen button to MenuBar directory
