+++ src/components/MenuBar/ToggleFullScreenButton/ToggleFullScreenButton.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { render, fireEvent } from '@testing-library/react';
-import useFullScreenToggle from '../../hooks/useFullScreenToggle/useFullScreenToggle';
+import useFullScreenToggle from '../../../hooks/useFullScreenToggle/useFullScreenToggle';
 
 import ToggleFullscreenButton from './ToggleFullScreenButton';
